--- a/zstd/seqdec_amd64.go
+++ b/zstd/seqdec_amd64.go
@@ -30,22 +30,12 @@ const errorMatchLenTooBig = 2
 // sequenceDecs_decode implements the main loop of sequenceDecs in x86 asm.
 //
 // Please refer to seqdec_generic.go for the reference implementation.
+//go:noescape
 func sequenceDecs_decode_amd64(s *sequenceDecs, br *bitReader, ctx *decodeAsmContext) int
 
 // sequenceDecs_decode implements the main loop of sequenceDecs in x86 asm with BMI2 extensions.
+//go:noescape
 func sequenceDecs_decode_bmi2(s *sequenceDecs, br *bitReader, ctx *decodeAsmContext) int
-
-type sequenceDecs_decode_function = func(s *sequenceDecs, br *bitReader, ctx *decodeAsmContext) int
-
-var sequenceDecs_decode sequenceDecs_decode_function
-
-func init() {
-	if cpuinfo.HasBMI2() {
-		sequenceDecs_decode = sequenceDecs_decode_bmi2
-	} else {
-		sequenceDecs_decode = sequenceDecs_decode_amd64
-	}
-}
 
 // decode sequences from the stream without the provided history.
 func (s *sequenceDecs) decode(seqs []seqVals) error {
@@ -70,7 +60,12 @@ func (s *sequenceDecs) decode(seqs []seqVals) error {
 
 	s.seqSize = 0
 
-	errCode := sequenceDecs_decode(s, br, &ctx)
+	var errCode int
+	if cpuinfo.HasBMI2() {
+		errCode = sequenceDecs_decode_bmi2(s, br, &ctx)
+	} else {
+		errCode = sequenceDecs_decode_amd64(s, br, &ctx)
+	}
 	if errCode != 0 {
 		i := len(seqs) - ctx.iteration
 		switch errCode {